Remove AMP from contribution to perf results
David Strauss suggested that we remove AMP from the potential
contributors to the results Rudy presented on performance improvements
in Google Search.  This is partly because AMP is unlikely now to
contribute to pre-render (only pre-fetch) and because it is unclear what
the status of that is.  This is all tied up in work with iframes and
portals in a way that is hard to predict.